Removed the contract of passing in a initial state to the undoredo reducer.
This cannot be enforced as the user will most likely create a store with just reducers
and that is a valid use case for creating a redux store.

The initial state is inferred from the previous reducers as the previous reducers
always return a default value (in our case the default value of the entire store)
which we can safely use.